Move test to test_shared_tasks
